Fix unused and raw type warnings

+++ contrib/labs/src/main/java/org/neo4j/graphalgo/results/VirtualNode.java
@@ -162,7 +162,7 @@ public class VirtualNode implements Node {
 
     @Override
     public int getDegree() {
-        return (int) rels.size();
+        return rels.size();
     }
 
     @Override
